scraper/requirements.txt
psycopg2-binary==2.9.9
playwright==1.40.0
python-dotenv==1.0.0
